Time: 4 ms (42.62%) | Memory: 69.5 MB (79.55%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -15,6 +15,6 @@ class Solution {
             max=Math.max(max,sum);
         }
 
-        return (double) max/k;
+        return  max/k;
     }
 }
